# specs/006-job-cancellation/openapi.yaml
openapi: 3.1.0
info:
  title: Retriva Job Cancellation API
  version: "0.9.0"
  description: >
    Job tracking and cooperative cancellation for asynchronous ingestion jobs.
    All ingestion endpoints return a job_id that can be used to poll status
    and request cancellation.

paths:
  /api/v1/jobs:
    get:
      summary: List all ingestion jobs
      operationId: listJobs
      description: >
        Returns all tracked ingestion jobs with their current status.
        Jobs are stored in-memory and lost on restart.
      responses:
        "200":
          description: List of jobs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/JobResponse"

  /api/v1/jobs/{job_id}:
    get:
      summary: Get job status
      operationId: getJob
      parameters:
        - name: job_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Job details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobResponse"
        "404":
          description: Job not found

  /api/v1/jobs/{job_id}/cancel:
    post:
      summary: Cancel a running or pending job
      operationId: cancelJob
      description: >
        Requests cooperative cancellation. The background worker will stop
        at the next batch boundary. Completed/failed jobs cannot be cancelled
        (returns 409).
      parameters:
        - name: job_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "202":
          description: Cancellation requested
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobResponse"
        "200":
          description: Already cancelling or cancelled (idempotent)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobResponse"
        "404":
          description: Job not found
        "409":
          description: Job already completed or failed (cannot cancel)

components:
  schemas:
    JobResponse:
      type: object
      required: [job_id, status, source, job_type, created_at, updated_at]
      properties:
        job_id:
          type: string
          example: "550e8400-e29b-41d4-a716-446655440000"
        status:
          type: string
          enum: [pending, running, completed, failed, cancelling, cancelled]
        source:
          type: string
          description: Source path from the ingestion request
        job_type:
          type: string
          enum: [html, text, image, chunks]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        error:
          type: string
          nullable: true
          description: Error message if status is 'failed'

    IngestResponse:
      type: object
      properties:
        status:
          type: string
        message:
          type: string
        job_id:
          type: string
          nullable: true
          description: >
            Job identifier for tracking. Optional for backward compatibility —
            existing clients that don't use job tracking are unaffected.