# src/retriva/ingestion_api/schemas.py
# Copyright (C) 2026 Andrea Marson (am.dev.75@gmail.com)

from pydantic import BaseModel, Field
from typing import List, Optional
from retriva.domain.models import Chunk

class HtmlIngestRequest(BaseModel):
    source_path: str = Field(..., description="Canonical URL or source path.")
    page_title: str = Field("", description="Title of the page.")
    html_content: str = Field(..., description="Raw HTML content.")
    origin_file_path: str = Field("", description="Original filesystem path of the HTML file (for resolving relative image paths).")

class TextIngestRequest(BaseModel):
    source_path: str = Field(..., description="Canonical URL or source path.")
    page_title: str = Field("", description="Title of the page.")
    content_text: str = Field(..., description="Plain text to chunk.")

class ImageIngestRequest(BaseModel):
    source_path: str = Field(..., description="Canonical URL or source path.")
    page_title: str = Field("", description="Title of the page.")
    file_path: str = Field(..., description="Absolute path to the image file on disk.")

class ChunkIngestRequest(BaseModel):
    chunks: List[Chunk] = Field(..., description="List of completely raw Chunk objects.")

class IngestResponse(BaseModel):
    status: str
    message: str
    job_id: Optional[str] = None

class JobResponse(BaseModel):
    job_id: str
    status: str
    source: str
    job_type: str
    created_at: str
    updated_at: str
    error: Optional[str] = None